seconnd

+++ module_2/lesson_10/src/solution_1.py
@@ -1,10 +1,10 @@
 def finish_price(price, amount):
     if amount >= 10 and amount < 20:
-        return price*0.9 
+        return int(price*0.9) 
     elif amount >=20:
-        return price*0.8
+        return int(price*0.8)
     else:
-        return price
+        return int(price)
 print(f"Итоговая сумма : {finish_price(100, 5)}") 
 print(f"Итоговая сумма : {finish_price(200, 10)}")
 print(f"Итоговая сумма : {finish_price(150, 20)}")
